+++ Jenkinsfile.groovy
@@ -1,20 +1,8 @@
-// Groovy
-//
-// Copyright 2021 MicroEJ Corp. All rights reserved.
-// This library is provided in source code for use, modification and test, subject to license terms.
-// Any modification of the source code will break MicroEJ Corp. warranties on the whole library.
-//
-//Global variables
-            buildWithAnt{
-              PROJECT_ID = "1545"			//The gitlab project ID. It can be found in the gitlab project interface.
-            
-              //MODULE_DIR = ""				// The name of the project into which your MODULE_FILENAME is. Default value is : "" . So leave blank if your MODULE_FILENAME is in the root of your gitlab project.
-              //MODULE_FILENAME= "module.ivy"				// The module filename. Default value is : module.ivy. Set it if yours is different.
-              //LABEL = ""					// The label where build must be run. Default value is : public. Set it if yours is different.
-              
-              //ANT_VERSION = ""				// The version of ant by Microej. Default value is : MMM MicroEJ 5.1.0. Set it if yours is different.
-              //ANT_PROPERTIES = ''				// The properties run with ant. Default value is already set. Add specific properties if necessary. As ('-DnameOfProperties=valueofProperty').
-              //ANT_VERBOSE = ""				// Default value is : false. Set true if you want the verbose mode.
-            
-              //PATTERN_JUNIT = ''				// The pattern Junit for test report. This default value is : '**/target~/test/xml/**/*test-report.xml, **/target~/test/xml/**/*Test.xml'. Set it if you want to use different one.
-            }
+// Copyright 2021-2022 MicroEJ Corp. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be found with this software.
+
+buildWithMMM {
+  MODULE_FILENAME= 'module.ivy'
+  DOCKER_IMAGE = 'artifactory.cross:18084/microej/sdk-cco:latest'
+  ANT_PROPERTIES = '-b=${IVY_SETTINGS_MODULES_REPO_PATH}'
+}