+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2022-06-30
+
+### Changed
+
+- Change license terms to MicroEJ Corp. BSD-style license.
+
+### Fixed
+
+- Fix application time that was not set properly in `LLMJVM_IMPL_setApplicationTime()`.
+- Fix invalid creation of ThreadX timer and semaphore in `LLMJVM_IMPL_initialize()`.
+- Fix ThreadX timer that was configured as a periodic timer instead of one-shot timer in `LLMJVM_IMPL_scheduleRequest()`.
+- Fix MicroEJ Time component that was not initialized.
+- Fix semaphore's initialization that caused the first call to `LLMJVM_IMPL_idleVM()` to be non-blocking.
+
 ## [1.0.0] - 2021-06-03
 
 ### Added
@@ -13,6 +27,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
-_Copyright 2021 MicroEJ Corp. All rights reserved._  
-_This library is provided in source code for use, modification and test, subject to license terms._  
-_Any modification of the source code will break MicroEJ Corp. warranties on the whole library._  
+_Copyright 2021-2022 MicroEJ Corp. All rights reserved._
+_Use of this source code is governed by a BSD-style license that can be found with this software._
+